+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-### v1.0.5
-- reset properties instead of removing attributes on reset
+### v3.0.0
+- Update source code to support Veams 5 framework
+
+### v2.0.0
+- make module compatible with VeamsQuery
+- rename option `showLoader` to `loadingClass`
+- add options `successClass` and `errorClass`
 
 ### v1.0.4
 - added global module registry